+++ kernel-rs/src/sysfile.rs
@@ -32,9 +32,9 @@ impl File {
         let mut p: *mut Proc = myproc();
         for fd in 0..NOFILE {
             // user pointer to struct stat
-            if (*p).ofile[fd as usize].is_null() {
-                (*p).ofile[fd as usize] = self;
-                return fd;
+            if (*p).ofile[fd].is_null() {
+                (*p).ofile[fd] = self;
+                return fd as i32;
             }
         }
         -1
@@ -49,7 +49,7 @@ unsafe fn argfd(n: i32, pfd: *mut i32, pf: *mut *mut File) -> i32 {
     if argint(n, &mut fd) < 0 {
         return -1;
     }
-    if fd < 0 || fd >= NOFILE || {
+    if fd < 0 || fd >= NOFILE as i32 || {
         f = (*myproc()).ofile[fd as usize];
         f.is_null()
     } {
@@ -123,9 +123,11 @@ pub unsafe fn sys_fstat() -> usize {
 /// Create the path new as a link to the same inode as old.
 pub unsafe fn sys_link() -> usize {
     let mut name: [u8; DIRSIZ] = [0; DIRSIZ];
-    let mut new: [u8; MAXPATH as usize] = [0; MAXPATH as usize];
-    let mut old: [u8; MAXPATH as usize] = [0; MAXPATH as usize];
-    if argstr(0, old.as_mut_ptr(), MAXPATH) < 0 || argstr(1, new.as_mut_ptr(), MAXPATH) < 0 {
+    let mut new: [u8; MAXPATH as usize] = [0; MAXPATH];
+    let mut old: [u8; MAXPATH as usize] = [0; MAXPATH];
+    if argstr(0, old.as_mut_ptr(), MAXPATH as i32) < 0
+        || argstr(1, new.as_mut_ptr(), MAXPATH as i32) < 0
+    {
         return usize::MAX;
     }
     begin_op();
@@ -189,9 +191,9 @@ unsafe fn isdirempty(dp: *mut Inode) -> i32 {
 pub unsafe fn sys_unlink() -> usize {
     let mut de: Dirent = Default::default();
     let mut name: [u8; DIRSIZ] = [0; DIRSIZ];
-    let mut path: [u8; MAXPATH as usize] = [0; MAXPATH as usize];
+    let mut path: [u8; MAXPATH] = [0; MAXPATH];
     let mut off: u32 = 0;
-    if argstr(0, path.as_mut_ptr(), MAXPATH) < 0 {
+    if argstr(0, path.as_mut_ptr(), MAXPATH as i32) < 0 {
         return usize::MAX;
     }
     begin_op();
@@ -292,11 +294,11 @@ unsafe fn create(path: *mut u8, typ: i16, major: i16, minor: i16) -> *mut Inode 
 }
 
 pub unsafe fn sys_open() -> usize {
-    let mut path: [u8; MAXPATH as usize] = [0; MAXPATH as usize];
+    let mut path: [u8; MAXPATH] = [0; MAXPATH];
     let mut fd: i32 = 0;
     let mut omode: i32 = 0;
     let ip: *mut Inode;
-    let n: i32 = argstr(0, path.as_mut_ptr(), MAXPATH);
+    let n: i32 = argstr(0, path.as_mut_ptr(), MAXPATH as i32);
     if n < 0 || argint(1, &mut omode) < 0 {
         return usize::MAX;
     }
@@ -354,10 +356,10 @@ pub unsafe fn sys_open() -> usize {
 }
 
 pub unsafe fn sys_mkdir() -> usize {
-    let mut path: [u8; MAXPATH as usize] = [0; MAXPATH as usize];
+    let mut path: [u8; MAXPATH] = [0; MAXPATH];
     let mut ip: *mut Inode = ptr::null_mut();
     begin_op();
-    if argstr(0, path.as_mut_ptr(), MAXPATH) < 0 || {
+    if argstr(0, path.as_mut_ptr(), MAXPATH as i32) < 0 || {
         ip = create(path.as_mut_ptr(), T_DIR as i16, 0, 0);
         ip.is_null()
     } {
@@ -371,11 +373,11 @@ pub unsafe fn sys_mkdir() -> usize {
 
 pub unsafe fn sys_mknod() -> usize {
     let mut ip: *mut Inode = ptr::null_mut();
-    let mut path: [u8; MAXPATH as usize] = [0; MAXPATH as usize];
+    let mut path: [u8; MAXPATH] = [0; MAXPATH];
     let mut major: i32 = 0;
     let mut minor: i32 = 0;
     begin_op();
-    if argstr(0, path.as_mut_ptr(), MAXPATH) < 0
+    if argstr(0, path.as_mut_ptr(), MAXPATH as i32) < 0
         || argint(1, &mut major) < 0
         || argint(2, &mut minor) < 0
         || {
@@ -397,11 +399,11 @@ pub unsafe fn sys_mknod() -> usize {
 }
 
 pub unsafe fn sys_chdir() -> usize {
-    let mut path: [u8; MAXPATH as usize] = [0; MAXPATH as usize];
+    let mut path: [u8; MAXPATH] = [0; MAXPATH];
     let mut ip: *mut Inode = ptr::null_mut();
     let mut p: *mut Proc = myproc();
     begin_op();
-    if argstr(0, path.as_mut_ptr(), MAXPATH) < 0 || {
+    if argstr(0, path.as_mut_ptr(), MAXPATH as i32) < 0 || {
         ip = namei(path.as_mut_ptr());
         ip.is_null()
     } {
@@ -423,12 +425,12 @@ pub unsafe fn sys_chdir() -> usize {
 
 pub unsafe fn sys_exec() -> usize {
     let current_block: usize;
-    let mut path: [u8; MAXPATH as usize] = [0; MAXPATH as usize];
+    let mut path: [u8; MAXPATH] = [0; MAXPATH];
     let mut argv: [*mut u8; MAXARG] = [ptr::null_mut(); MAXARG];
     let mut i: i32 = 0;
     let mut uargv: usize = 0;
     let mut uarg: usize = 0;
-    if argstr(0, path.as_mut_ptr(), MAXPATH) < 0 || argaddr(1, &mut uargv) < 0 {
+    if argstr(0, path.as_mut_ptr(), MAXPATH as i32) < 0 || argaddr(1, &mut uargv) < 0 {
         return usize::MAX;
     }
     ptr::write_bytes(argv.as_mut_ptr(), 0, 1);
@@ -457,7 +459,7 @@ pub unsafe fn sys_exec() -> usize {
             if argv[i as usize].is_null() {
                 panic(b"sys_exec kalloc\x00" as *const u8 as *mut u8);
             }
-            if fetchstr(uarg, argv[i as usize], PGSIZE) < 0 {
+            if fetchstr(uarg, argv[i as usize], PGSIZE as i32) < 0 {
                 current_block = 12646643519710607562;
                 break;
             }
